判断文件是否存在

+++ build.sh
@@ -3,40 +3,48 @@ path=$1
 
 if [ ! -n "$path" ] ;then
      echo "please input nginx install path"
-     exit
+     exit 1
 fi
 git submodule update --init --recursive
 
-cp -r ./nginx  ${path}/nginx/
+cp -r ./nginx  ${path}nginx
+if [[ ! -f "LuaJIT.tar.gz"  ]]; then
+    wget http://luajit.org/download/LuaJIT-2.0.5.tar.gz -O LuaJIT.tar.gz
 
-wget http://luajit.org/download/LuaJIT-2.0.5.tar.gz -O LuaJIT.tar.gz
+fi
 tar zxf LuaJIT.tar.gz
-
 cd LuaJIT-2.0.5
-make PREFIX=${path}/nginx/luajit
-make install PREFIX=${path}/nginx/luajit
+make PREFIX=${path}nginx/luajit
+make install PREFIX=${path}nginx/luajit
 cd ..
 
-export LUAJIT_LIB=${path}/nginx/luajit/lib
-export LUAJIT_INC=${path}/nginx/luajit/include//luajit-2.0/
-
-wget https://ftp.pcre.org/pub/pcre/pcre-8.00.tar.bz2 -O pcre.tar.bz2
+export LUAJIT_LIB=${path}nginx/luajit/lib
+export LUAJIT_INC=${path}nginx/luajit/include//luajit-2.0/
+if [[ ! -f "pcre.tar.bz2" ]]; then
+    wget https://ftp.pcre.org/pub/pcre/pcre-8.00.tar.bz2 -O pcre.tar.bz2
+fi
 tar -xvf pcre.tar.bz2
-
-wget https://bitbucket.org/nginx-goodies/nginx-sticky-module-ng/get/master.tar.gz -O nginx-sticky-module.tar.gz
+if [[ ! -f "nginx-sticky-module.tar.gz" ]]; then
+    wget https://bitbucket.org/nginx-goodies/nginx-sticky-module-ng/get/master.tar.gz -O nginx-sticky-module.tar.gz
+fi
 tar -zxvf nginx-sticky-module.tar.gz
 
-wget https://www.openssl.org/source/openssl-1.1.1-pre1.tar.gz -O openssl.tar.gz
+if [[ ! -f "openssl.tar.gz" ]]; then
+    wget https://www.openssl.org/source/openssl-1.1.1-pre1.tar.gz -O openssl.tar.gz
+fi
+
 tar -xvf openssl.tar.gz
 mv nginx-goodies-nginx-sticky-module-ng-08a395c66e42  nginx-sticky-module
 
-wget http://nginx.org/download/nginx-1.13.9.tar.gz -O nginx.tar.gz
+if [[ ! -f "nginx.tar.gz" ]]; then
+    wget http://nginx.org/download/nginx-1.13.9.tar.gz -O nginx.tar.gz
+fi
 tar -xvf nginx.tar.gz
 
 
 cd nginx-1.13.9
 
-./configure --prefix=${path}/nginx \
+./configure --prefix=${path}nginx \
             --with-http_ssl_module \
             --with-http_gzip_static_module \
             --with-http_stub_status_module \
